chore(twilio): refactor code structure
