+++ integrations/twilio/definitions/index.ts
@@ -1,0 +1,4 @@
+export { channels } from './channels'
+export { configuration } from './configuration'
+export { entities } from './entities'
+export { user } from './user'
